Remove GenericVersionStrategy concept as it was superseded by templates
