+++ src/output/yaml.ts
@@ -2,7 +2,7 @@ import YAML from "yaml";
 
 import type { OutputPrinter } from ".";
 import type { Config } from "../config/types";
-import type { VersionResult } from "../versionResolver";
+import type { VersionResult } from "../version/versionResolver";
 
 export class YamlOutputPrinter implements OutputPrinter {
   printResult(_config: Config, versionResult: VersionResult) {
