+++ src/version/versionResolver.ts
@@ -1,12 +1,12 @@
 import semver, { SemVer } from "semver";
 import slug from "slug";
 
-import type { Config } from "./config/types";
-import { logger } from "./logging";
-import type { Platform } from "./platform";
-import { getCommitDateTime, listTags, listTagsBeforeCommit } from "./util/git";
-import { semVerVersionString } from "./util/semVer";
-import type { VersionStrategy } from "./version";
+import type { Config } from "../config/types";
+import { logger } from "../logging";
+import type { Platform } from "../platform";
+import { getCommitDateTime, listTags, listTagsBeforeCommit } from "../util/git";
+import { semVerVersionString } from "../util/semVer";
+import type { VersionStrategy } from "./versionStrategy";
 
 const versionResolveLogger = logger.childLogger("version-resolver");
